requirements.txt
adal==1.2.7
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
ansiwrap==0.8.4
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-papermill==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==3.0.8
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-datalake-store==0.0.52
azure-storage-blob==12.16.0
Babel==2.12.1
backcall==0.2.0
black==23.3.0
blinker==1.6.2
blis==0.7.9
boto3==1.26.119
botocore==1.29.119
cachelib==0.9.0
cachetools==5.3.0
catalogue==2.0.8
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
confection==0.0.4
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==40.0.2
cymem==2.0.7
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2023.4.0
gcsfs==2023.4.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
inflection==0.5.1
ipython==8.12.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
langcodes==3.3.0
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.7.3
nbformat==5.8.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
papermill==2.4.0
parso==0.8.3
pathspec==0.9.0
pathy==0.10.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
preshed==3.0.8
prison==0.2.1
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rich==13.3.4
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
scrapbook==0.5.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.0
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
textwrap3==0.9.2
thinc==8.1.9
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.1
